pyproject.toml
[project]
name = "ynk-modelo"
version = "0.1.0"
description = "Generador de estados de resultados y simulador EBITDA para tiendas Yáneken"
authors = [{ name = "YNK" }]
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "openpyxl>=3.1",
    "python-dotenv>=1.0",
    "flask>=3.0",
    "flask-login>=0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[project.scripts]
ynk-eerr = "ynk_modelo.cli.main:main"
ynk-auto = "ynk_modelo.cli.auto_regenerate:main"
ynk-server = "ynk_modelo.cli.server:main"
ynk-server-auth = "ynk_modelo.cli.flask_server:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["ynk_modelo*"]
